Change ActiveModel's ImmutableString#type to match its registry entry.
ActiveModel already 'manually' registers the ImmutableString type under
the name ":immutable_string" as you'd expect.

Currently, if a user tries to register ImmutableString in order to be
able to write "attribute :status, :immutable_string" in a model, the
registry ends up with two types keyed to the name ':string'; due to the
order in which the types are registered, this happens to not interfere
with "attribute :mutable, :string".

This change removes the inconsistency between registrar and registree.